fixed faulty command crafting to solve issue 13016. Command was wrongfully formatted for git. Faulty Example: git checkout --progress origin/[Branch Name] -b [Branch Name] --recurse-submodules --
Working crafted command:
git checkout --progress -b [Branch Name OR Origin Branch Name] --recurse-submodules --

The Checkout now is rightfully formatted and pulls submodules aswell. It was an Formatting Issue it was only checked if the name is a remote branch and the according -b [Branch Name] appended. A Check for the Local Branch was missing. therefore Local and Remote version of branchname could be appended to the args which caused the error in the first place.

Also i am ver yunsure if the -- at the end are really necessary i dont see them at the end of any command elsewhere in the application, its up to the reviewer to test it and remove if unnecessary. Good Day Enjoy the simple fix for a Bug which was in the Application for at least 6 years.